requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
biopython>=1.79

# Bioinformatics tools
pysam>=0.19.0
pysamstats>=1.1.2

# Feature extraction
numba>=0.56.0
joblib>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Hashing and indexing
mmh3>=3.0.0
datasketch>=1.5.0

# CLI and configuration
click>=8.0.0
pyyaml>=6.0
tqdm>=4.62.0

# Logging and monitoring
colorlog>=6.6.0

# Quality control wrappers
subprocess32>=3.5.4; python_version < "3"

# Testing dependencies (optional)
pytest>=6.0.0
pytest-cov>=3.0.0

# Documentation dependencies (optional)
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Development dependencies (optional)
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0